+++ Changelog.md
@@ -1,5 +1,7 @@
 # Next
 
+- Adds `response` computed property to `Error` type, which yields a Response object if available.
+
 # 6.1.3
 
 - Updated to ReactiveCocoa 4.0 final.