+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Adds `response` computed property to `Error` type, which yields a Response object if available.
 - Added URLEncodedInURL to ParameterEncoding.
+- Adds convenience `endpointByAdding` method.
 
 # 6.1.3